+++ CHANGELOG.md
@@ -86,6 +86,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Fix block playout random seeds
   - Different blocks within a single playout will now correctly use different random seeds (shuffles)
   - Erasing block playout history will also generate new random seeds for the playout
+- Fix building playouts that use mid-roll filler and have content without chapter markers
+  - When this happens, mid-roll will be treated as post-roll
 
 ### Changed
 - Use ffmpeg 7 in all docker images